requirements.txt
Django==3.2.4
pytz==2021.1
psycopg2-binary>=2.8
mysqlclient>=2.0
django-debug-toolbar==3.8.1
djangorestframework==3.12.4